refactor: move a couple more files
